Changed Song ID system to use CRC-32 checksum of song file.
This'll provide globally unique IDs for each song that will only collide if two songs have the exact same song file.

+++ src/persistence/LocalSong.java
@@ -23,7 +23,7 @@ public class LocalSong implements Song {
    private String album;
    private long duration;
    private Library lib;
-   private int id;
+   private long id;
    private File path;
 
    /**
@@ -36,7 +36,7 @@ public class LocalSong implements Song {
     * @param path     the path to the file containing the song
     * @param id       the ID number of the song
     */
-   protected LocalSong(String title, String artist, String album, long duration, File path, int id) {
+   protected LocalSong(String title, String artist, String album, long duration, File path, long id) {
       this.title = title;
       this.artist = artist;
       this.album = album;
@@ -59,7 +59,7 @@ public class LocalSong implements Song {
       String artist = null;
       long duration = 0;
       String path = null;
-      int id = 0;
+      long id = 0;
 
       JsonToken token = parser.currentToken();
       while (token != JsonToken.END_OBJECT) {
@@ -94,7 +94,7 @@ public class LocalSong implements Song {
     *
     * @return song ID number
     */
-   public int getId() {
+   public long getId() {
       return this.id;
    }
